test: move NoOpFlywayConfiguration into dhis-support-test
so we do not include it into the dhis.war and risk using it in production
